⛏️ :: Delete 추가

+++ src/main/java/com/onpost/domain/service/MemberService.java
@@ -1,10 +1,6 @@
 package com.onpost.domain.service;
 
-import com.onpost.domain.dto.member.FollowDto;
-import com.onpost.domain.dto.member.MemberRequest;
-import com.onpost.domain.dto.member.MemberResponse;
-import com.onpost.domain.dto.member.MemberView;
-import com.onpost.domain.entity.Image;
+import com.onpost.domain.dto.member.*;
 import com.onpost.domain.entity.Post;
 import com.onpost.domain.entity.member.Member;
 import com.onpost.domain.repository.MemberQueryRepository;
@@ -13,10 +9,15 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
+@Transactional(rollbackFor = {Exception.class})
 public class MemberService {
 
     private final MemberQueryRepository memberQueryRepository;
@@ -24,7 +25,7 @@ public class MemberService {
     private final PostQueryRepository postQueryRepository;
 
     public MemberView editMember(MemberRequest memberRequest) {
-        Member member = memberQueryRepository.getCurrentMember(memberRequest.getId());
+        Member member = memberQueryRepository.findMember(memberRequest.getId());
 
         if(memberRequest.getName() != null) {
             member.setName(memberRequest.getName());
@@ -46,40 +47,50 @@ public class MemberService {
     }
 
     public MemberResponse showMember(Long id) {
-        return new MemberResponse(memberQueryRepository.getMemberAll(id));
+        return new MemberResponse(memberQueryRepository.findOneWithAll(id));
     }
 
-    public void followMember(FollowDto followDto) {
-        Member me = memberQueryRepository.getFollowRelation(followDto.getId());
-        Member follow = memberQueryRepository.getFollowRelation(followDto.getFollowId());
+    public void followMember(FollowDto followDto, boolean positive) {
+        Member me = memberQueryRepository.findOneWithFollow(followDto.getId());
+        Member follow = memberQueryRepository.findOneWithFollow(followDto.getFollowId());
 
-        me.followMe(follow);
+        if(positive) {
+            follow.followMe(me);
+        } else {
+            follow.unfollowMe(me);
+        }
 
         memberQueryRepository.save(me);
         memberQueryRepository.save(follow);
     }
 
+    public FollowResponse followList(Long id) {
+        Member member = memberQueryRepository.findOneWithFollow(id);
+        return FollowResponse.builder()
+                .follower(member.getFollower().stream().map(MemberView::new).collect(Collectors.toList()))
+                .following(member.getFollowing().stream().map(MemberView::new).collect(Collectors.toList()))
+                .build();
+    }
+
     public void deleteMember(Long id) {
-        Member member = memberQueryRepository.getMemberAll(id);
-        for(Member m : member.getFollowing()) {
-            m.unfollowMe(member);
+        Member member = memberQueryRepository.findOneWithAll(id);
+
+        member.getFollower().forEach(m -> m.followMe(member));
+        member.getFollower().forEach(member::unfollowMe);
+
+        List<Post> posts = List.copyOf(member.getMakePost());
+        posts.forEach(postQueryRepository::delete);
+
+        if(member.getProfile() != null) {
+            imageService.deletePath(member.getProfile());
         }
 
-        for(Member m : member.getFollower()) {
-            member.unfollowMe(m);
-        }
-
-        for(Post p : member.getMakePost()) {
-            postQueryRepository.delete(p);
-        }
-
-        imageService.deletePath(member.getProfile());
         memberQueryRepository.delete(member);
     }
 
     public MemberView infoMember() {
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
-        Member member = memberQueryRepository.getMemberByEmail(email);
+        Member member = memberQueryRepository.findOneByEmail(email);
         return new MemberView(member);
     }
 }
